changed paddle pos message to float

+++ Assets/Scripts/PaletteScript.cs
@@ -95,7 +95,7 @@ public class PaletteScript : MonoBehaviour
         //Send OSC message
         var message = new OSCMessage(string.Format("/paddle/pos/y", input));
         // Populate values.
-        message.AddValue(OSCValue.Int((int)transform.position.y));
+        message.AddValue(OSCValue.Float(transform.position.y));
         _transmitter.Send(message);
